refactor(web): extract use-event-bus-init Effect 2 into assistant/sse-service
The SSE connection and bounce policy lift out of `useEventBusInit`'s
Effect 2 into `assistant/sse-service.ts`, a non-React service that
parallels `lifecycleService`. The hook becomes a thin React adapter:
two effects, neither doing real work — Effect 1 wires the runtime
event-sources at mount, Effect 2 calls `sseService.attach` when the
active assistant changes and the returned detach on unmount or
input change.

`sseService.attach(bus, assistantId)` opens the daemon's `/v1/events`
stream, republishes envelopes as `sse.event`, publishes `sse.opened`
/ `sse.closed` lifecycle signals, and drives the bounce policy from
`app.hidden` / `app.resume` / `power.*` / `reachability.retry-requested`.
Each call creates a fresh per-assistant closure — independent dedup
timestamps, no leaked `current` reference across attachments.

A new narrow type `EventBusSubscriber = Pick<EventBusActions,
"subscribe">` parallels `EventBusPublisher` so the service signature
expresses its dual role (publish AND subscribe) at the type level.
Least-authority — same convention as the source helpers.

Tests follow the surface: 23 SSE-behavior tests move into
`assistant/sse-service.test.ts` and exercise the service directly
with an in-memory bus stub (no `renderHook`). The hook test slims to
6 cases asserting only the React-adapter contract (attach on active,
detach on unmount/changed-id/inactive). The hook test mocks the
upstream `lifecycle-service` / `stream-transport` modules only so
the import chain doesn't touch the daemon at module-eval time, then
`spyOn(sseService, "attach")` patches the real object — avoids
the `mock.module` process-global leak that would shadow
`sse-service.test.ts`'s real implementation.

`EVENT_BUS.md` updated: the "Where it lives" table reflects the
three-file model (bus / runtime sources / sse-service), the producer
section names `sse-service` for SSE-derived events, and the "Don't
do this" rules point at the right files.

`useEventBusInit` goes from 234 to 62 lines; the orchestration logic
that was there now lives in 200 lines of `sse-service.ts` with
explicit per-attachment state, no `useEffect` cleanup re-entrancy
surface.
